Initial shared workflow
As a test, this is a shared workflow for publishing an NPM package to
npmjs using pnpm. The action requires that you give it the version tag
to release under, which can optionally start with a v, and that the
package.json have a {{VERSION}} string in the version field.

The action will update the package.json and then publish.